+++ source/extensions/omni.isaac.core/omni/isaac/core/utils/semantics.py
@@ -8,6 +8,7 @@
 #
 
 from pxr import Semantics, Usd
+from typing import List, Tuple, Dict
 
 
 def add_update_semantics(prim: Usd.Prim, semantic_label: str, type_label: str = "class", suffix="") -> None:
@@ -65,3 +66,25 @@ def remove_all_semantics(prim: Usd.Prim, recursive: bool = False) -> None:
             remove_semantics(p)
     else:
         remove_semantics(prim)
+
+
+def get_semantics(prim: Usd.Prim) -> Dict[str, Tuple[str, str]]:
+    """Returns semantics that are applied to a prim
+    
+    Args:
+        prim (Usd.Prim): Prim to return semantics for
+
+    Returns:
+        Dict[str, Tuple[str,str]]: Dictionary containing the name of the applied semantic, and the type and data associated with that semantic. 
+    """
+    result = {}
+    for prop in prim.GetProperties():
+        is_semantic = Semantics.SemanticsAPI.IsSemanticsAPIPath(prop.GetPath())
+        if is_semantic:
+            name = prop.SplitName()[1]
+            sem = Semantics.SemanticsAPI.Get(prim, name)
+
+            typeAttr = sem.GetSemanticTypeAttr()
+            dataAttr = sem.GetSemanticDataAttr()
+            result[name] = (typeAttr.Get(), dataAttr.Get())
+    return result